Conflicts:
	CM4/Debug/System/System.o
	CM4/Debug/System/System.su
	CM4/Debug/XploreAvionics_CM4.bin
	CM4/Debug/XploreAvionics_CM4.elf
	CM4/Debug/XploreAvionics_CM4.list
	CM4/Debug/XploreAvionics_CM4.map
	System/Sensors/ADC16.cpp
	System/Sensors/ADC16.h
	System/Sensors/ADC24.cpp
	System/Sensors/ADC24.h
	System/Sensors/DataStructures.h
	System/System.cpp

+++ System/Sensors/DataStructures.h
@@ -66,10 +66,10 @@ struct BaroData {
 };
 
 struct ScienceData {
-	uint32_t mass;
+	float mass;
 
 	char* toString(char* buffer) {
-		sprintf(buffer, "Mass(kg): %lu", mass);
+		sprintf(buffer, "Mass(kg): %f", mass);
 		return buffer;
 	}
